вынес класс PageViewController в отдельный файл, переместил онбординг в отдельную папку Scenes
